add new routes for directory pages
