+++ app/controllers/directory/list/detail.js
@@ -1,21 +1,24 @@
 import Ember from 'ember';
+import config from '../../../config/environment';
 
-export default Ember.Component.extend({
+export default Ember.Controller.extend({
+  session: Ember.inject.service('session'),
   checkRoles: false,
+  user: Ember.computed.alias('model.user'),
+
   isEvaluator: Ember.computed('checkRoles', function () {
     return this.get('user.Roles').filter(x => x.get('name') === 'CORRECTOR').length !== 0;
   }),
+
   isStudent: Ember.computed('checkRoles', function () {
     return this.get('user.Roles').filter(x => x.get('name') === 'STUDENT').length !== 0;
   }),
+
   currentRoles: Ember.computed('user.Roles', 'user.Roles.[]', 'user.Roles.@each', function () {
     return this.get('user.Roles').map(x => x.get('name'));
   }),
 
-  init() {
-    this._super(...arguments);
-    this.set('newEmail', this.get('user.email'));
-  },
+  newEmail: Ember.computed.oneWay('user.email'),
 
   successfullPwd: false,
   successfullEmail: false,
@@ -42,12 +45,33 @@ export default Ember.Component.extend({
       return false;
     },
 
-    closeSettings() {
-      return this.sendAction('closeSettings');
-    },
-
     impersonateUser() {
-      return this.sendAction('impersonateUser');
+        const errorMessage = error => {
+          alert(error);
+        };
+        Ember.$.ajax({
+          type: "POST",
+          url: config.backendUrl + "/auth/impersonate",
+          data: JSON.stringify({ userId: +this.get('user.id') }),
+          beforeSend: (xhr) => { xhr.setRequestHeader('Authorization', `Bearer ${this.get('session.data.authenticated.token')}`); },
+          contentType: "application/json; charset=utf-8",
+          crossDomain: true,
+          dataType: "json",
+          success: (data) => {
+            if (data.token) {
+              this.set('session.data.authenticated.token', data.token);
+              this.get('session').trigger('authenticationSucceeded');
+              this.transitionToRoute('settings');
+            }
+          },
+          failure: errorMessage,
+          statusCode: {
+            500: errorMessage,
+            404: errorMessage,
+            403: errorMessage,
+          }
+        });
+        return false;
     },
 
     toggleRole(role) {
@@ -71,8 +95,4 @@ export default Ember.Component.extend({
     }
   },
 
-  willDestroyElement() {
-    this._super(...arguments);
-    this.get('user').rollbackAttributes();
-  }
 });
