+++ app/router.js
@@ -7,6 +7,20 @@ const Router = Ember.Router.extend({
 });
 
 Router.map(function() {
+  this.route('change-group');
+  this.route('demonstrator', function() {
+    this.route('detail', { path: "/:id" }, function() {
+      this.route('event', { path: "/event/:id" });
+    });
+  });
+  this.route('directory', function() {
+    this.route('new');
+    this.route('list', function() {
+      this.route('detail', { path: "/:id" });
+    });
+  });
+  this.route('entry-test');
+  this.route('error');
   this.route('evaluator', function() {
     this.route('to-fix', function() {
       this.route('deliverable', { path: "/:id" });
@@ -16,27 +30,18 @@ Router.map(function() {
       this.route('no-deliverable');
     });
   });
-  this.route('entry-test');
-  this.route('directory');
-  this.route('demonstrator', function() {
-    this.route('detail', { path: "/:id" }, function() {
-      this.route('event', { path: "/event/:id" });
-    });
-  });
-  this.route('semester');
+  this.route('event-templates');
   this.route('login');
+  this.route('login-shibboleth');
+  this.route('logout');
+  this.route('news');
   this.route('notification');
+  this.route('permission-denied');
+  this.route('semester');
   this.route('settings');
   this.route('sql');
   this.route('statistics');
   this.route('student');
-  this.route('logout');
-  this.route('error');
-  this.route('permission-denied');
-  this.route('news');
-  this.route('event-templates');
-  this.route('login-shibboleth');
-  this.route('change-group');
 });
 
 export default Router;
