+++ app/routes/directory/index.js
@@ -1,0 +1,8 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  beforeModel() {
+    this.replaceWith("directory.list");
+  }
+});
+
