+++ app/routes/directory/list.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+import AuthenticatedRouteMixin from 'ember-simple-auth/mixins/authenticated-route-mixin';
+import ErrorRouteMixin from '../../mixins/error-route';
+import AdminAuthenticatedRouteMixin from '../../mixins/admin-authenticated-route';
+
+export default Ember.Route.extend(AuthenticatedRouteMixin, ErrorRouteMixin, AdminAuthenticatedRouteMixin, {
+});
