+++ app/templates/directory/index.hbs
@@ -1,0 +1,1 @@
+{{outlet}}
